add endpoint playgrounds/<int:playground_id>/bookings/
